+++ gog.js
@@ -611,7 +611,7 @@ try {
   });
 } catch (error) {
   process.exitCode ||= 1;
-  log.fail(`Exception: ${error.message || error}`);
+  log.exception(error);
   if (cfg.debug) console.error(error);
   if (error.message && process.exitCode != 130) await notify(`gog failed: ${error.message.split('\n')[0]}`, { attachLatestScreenshot: true });
 } finally {